Switch the `tpm reset` and `tpm reset-thl` commands to use `trepctl reset -all -y`
git-svn-id: http://tungsten-replicator.googlecode.com/svn/trunk@2024 9f5e3340-5c9a-25cb-7c1c-c133a2e9e928

+++ builder/extra/tools/ruby-tpm/configure/commands/reset_cluster.rb
@@ -93,7 +93,8 @@ module ResetClusterDeploymentStep
       ConfigureCommitmentMethod.new("stop_replication_services", -1, 0),
       ConfigureCommitmentMethod.new("clear_dynamic_properties", 0, 0),
       ConfigureCommitmentMethod.new("rotate_logs", 0, 0),
-      ConfigureCommitmentMethod.new("start_replication_services", 1, ConfigureDeployment::FINAL_STEP_WEIGHT),
+      ConfigureCommitmentMethod.new("start_replication_services_offline", 0, 1),
+      ConfigureCommitmentMethod.new("reset_replication_services", 0, 2),
       ConfigureCommitmentMethod.new("wait_for_manager", 2, -1),
       ConfigureCommitmentMethod.new("set_original_policy", 4, 2),
       ConfigureCommitmentMethod.new("report_services", ConfigureDeployment::FINAL_GROUP_ID, ConfigureDeployment::FINAL_STEP_WEIGHT, false)
@@ -110,47 +111,52 @@ module ResetClusterDeploymentStep
           next
         end
         
-        ds = get_applier_datasource(rs_alias)
-        ds.drop_tungsten_schema(@config.getProperty([REPL_SERVICES, rs_alias, REPL_SVC_SCHEMA]))
-        
         thl_dir = @config.getProperty([REPL_SERVICES, rs_alias, REPL_LOG_DIR])
-        if get_additional_property(ResetClusterCommand::DELETE_LOGS) == "true"
-          debug("Remove all files in #{thl_dir}")
-          cmd_result("rm -f #{thl_dir}/*")
-        else
-          debug("Rename #{thl_dir}")
-          cmd_result("mv #{thl_dir} #{thl_dir}_#{get_additional_property(ResetClusterCommand::ARCHIVE_LOGS_SUFFIX)}")
-          mkdir_if_absent(thl_dir)
+        if File.exist?(thl_dir)
+          if get_additional_property(ResetClusterCommand::DELETE_LOGS) == "true"
+            debug("Remove all files in #{thl_dir}")
+            cmd_result("rm -f #{thl_dir}/*")
+          else
+            debug("Rename #{thl_dir}")
+            cmd_result("mv #{thl_dir} #{thl_dir}_#{get_additional_property(ResetClusterCommand::ARCHIVE_LOGS_SUFFIX)}")
+            mkdir_if_absent(thl_dir)
+          end
         end
         
         relay_dir = @config.getProperty([REPL_SERVICES, rs_alias, REPL_RELAY_LOG_DIR])
-        if get_additional_property(ResetClusterCommand::DELETE_LOGS) == "true"
-          debug("Remove all files in #{relay_dir}")
-          cmd_result("rm -f #{relay_dir}/*")
-        else
-          debug("Rename #{relay_dir}")
-          cmd_result("mv #{relay_dir} #{relay_dir}_#{get_additional_property(ResetClusterCommand::ARCHIVE_LOGS_SUFFIX)}")
-          mkdir_if_absent(relay_dir)
+        if File.exist?(relay_dir)
+          if get_additional_property(ResetClusterCommand::DELETE_LOGS) == "true"
+            debug("Remove all files in #{relay_dir}")
+            cmd_result("rm -f #{relay_dir}/*")
+          else
+            debug("Rename #{relay_dir}")
+            cmd_result("mv #{relay_dir} #{relay_dir}_#{get_additional_property(ResetClusterCommand::ARCHIVE_LOGS_SUFFIX)}")
+            mkdir_if_absent(relay_dir)
+          end
         end
         
         backup_dir = @config.getProperty([REPL_SERVICES, rs_alias, REPL_BACKUP_STORAGE_DIR])
         debug("Remove all files in #{backup_dir}")
         cmd_result("#{get_root_prefix()} rm -rf #{backup_dir}/*")
         
-        if get_additional_property(ResetClusterCommand::DELETE_LOGS) == "true"
-          debug("Remove all files in #{@config.getProperty(CURRENT_RELEASE_DIRECTORY)}/tungsten-replicator/log")
-          FileUtils.rm(Dir.glob("#{@config.getProperty(CURRENT_RELEASE_DIRECTORY)}/tungsten-replicator/log/*"))
-        else
-          debug("Rename files in #{@config.getProperty(CURRENT_RELEASE_DIRECTORY)}/tungsten-replicator/log")
-          cmd_result("cd #{@config.getProperty(CURRENT_RELEASE_DIRECTORY)}/tungsten-replicator/log; for i in *; do mv $i ${i}\"_#{get_additional_property(ResetClusterCommand::ARCHIVE_LOGS_SUFFIX)}\"; done")
+        if @config.getProperty(HOST_ENABLE_REPLICATOR) == "true"
+          if get_additional_property(ResetClusterCommand::DELETE_LOGS) == "true"
+            debug("Remove all files in #{@config.getProperty(CURRENT_RELEASE_DIRECTORY)}/tungsten-replicator/log")
+            FileUtils.rm(Dir.glob("#{@config.getProperty(CURRENT_RELEASE_DIRECTORY)}/tungsten-replicator/log/*"))
+          else
+            debug("Rename files in #{@config.getProperty(CURRENT_RELEASE_DIRECTORY)}/tungsten-replicator/log")
+            cmd_result("cd #{@config.getProperty(CURRENT_RELEASE_DIRECTORY)}/tungsten-replicator/log; for i in *; do mv $i ${i}\"_#{get_additional_property(ResetClusterCommand::ARCHIVE_LOGS_SUFFIX)}\"; done")
+          end
         end
         
-        if get_additional_property(ResetClusterCommand::DELETE_LOGS) == "true"
-          debug("Remove all files in #{@config.getProperty(CURRENT_RELEASE_DIRECTORY)}/tungsten-manager/log")
-          FileUtils.rm(Dir.glob("#{@config.getProperty(CURRENT_RELEASE_DIRECTORY)}/tungsten-manager/log/*"))
-        else
-          debug("Rename files in #{@config.getProperty(CURRENT_RELEASE_DIRECTORY)}/tungsten-manager/log")
-          cmd_result("cd #{@config.getProperty(CURRENT_RELEASE_DIRECTORY)}/tungsten-manager/log; for i in *; do mv $i ${i}\"_#{get_additional_property(ResetClusterCommand::ARCHIVE_LOGS_SUFFIX)}\"; done")
+        if @config.getProperty(HOST_ENABLE_MANAGER) == "true"
+          if get_additional_property(ResetClusterCommand::DELETE_LOGS) == "true"
+            debug("Remove all files in #{@config.getProperty(CURRENT_RELEASE_DIRECTORY)}/tungsten-manager/log")
+            FileUtils.rm(Dir.glob("#{@config.getProperty(CURRENT_RELEASE_DIRECTORY)}/tungsten-manager/log/*"))
+          else
+            debug("Rename files in #{@config.getProperty(CURRENT_RELEASE_DIRECTORY)}/tungsten-manager/log")
+            cmd_result("cd #{@config.getProperty(CURRENT_RELEASE_DIRECTORY)}/tungsten-manager/log; for i in *; do mv $i ${i}\"_#{get_additional_property(ResetClusterCommand::ARCHIVE_LOGS_SUFFIX)}\"; done")
+          end
         end
       }
     end
