+++ builder/extra/tools/ruby-tpm/configure/commands/reset_thl.rb
@@ -92,7 +92,8 @@ module ResetTHLDeploymentStep
       ConfigureCommitmentMethod.new("set_maintenance_policy", ConfigureDeployment::FIRST_GROUP_ID, 0),
       ConfigureCommitmentMethod.new("stop_replication_services", -1, 0),
       ConfigureCommitmentMethod.new("rotate_logs", 0, 0),
-      ConfigureCommitmentMethod.new("start_replication_services", 1, ConfigureDeployment::FINAL_STEP_WEIGHT),
+      ConfigureCommitmentMethod.new("start_replication_services_offline", 0, 1),
+      ConfigureCommitmentMethod.new("reset_replication_services", 0, 2),
       ConfigureCommitmentMethod.new("wait_for_manager", 2, -1),
       ConfigureCommitmentMethod.new("set_original_policy", 4, 2),
       ConfigureCommitmentMethod.new("report_services", ConfigureDeployment::FINAL_GROUP_ID, ConfigureDeployment::FINAL_STEP_WEIGHT, false)
@@ -109,29 +110,28 @@ module ResetTHLDeploymentStep
           next
         end
         
-        ds = get_applier_datasource(rs_alias)
-        if ds.is_a?(MySQLDatabasePlatform)
-          ds.run("set sql_log_bin=0;drop schema #{@config.getProperty([REPL_SERVICES, rs_alias, REPL_SVC_SCHEMA])}")
-        end
-        
         thl_dir = @config.getProperty([REPL_SERVICES, rs_alias, REPL_LOG_DIR])
-        if get_additional_property(ResetClusterCommand::DELETE_LOGS) == "true"
-          debug("Remove all files in #{thl_dir}")
-          cmd_result("rm -f #{thl_dir}/*")
-        else
-          debug("Rename #{thl_dir}")
-          cmd_result("mv #{thl_dir} #{thl_dir}_#{get_additional_property(ResetClusterCommand::ARCHIVE_LOGS_SUFFIX)}")
-          mkdir_if_absent(thl_dir)
+        if File.exist?(thl_dir)
+          if get_additional_property(ResetClusterCommand::DELETE_LOGS) == "true"
+            debug("Remove all files in #{thl_dir}")
+            cmd_result("rm -f #{thl_dir}/*")
+          else
+            debug("Rename #{thl_dir}")
+            cmd_result("mv #{thl_dir} #{thl_dir}_#{get_additional_property(ResetClusterCommand::ARCHIVE_LOGS_SUFFIX)}")
+            mkdir_if_absent(thl_dir)
+          end
         end
         
         relay_dir = @config.getProperty([REPL_SERVICES, rs_alias, REPL_RELAY_LOG_DIR])
-        if get_additional_property(ResetClusterCommand::DELETE_LOGS) == "true"
-          debug("Remove all files in #{relay_dir}")
-          cmd_result("rm -f #{relay_dir}/*")
-        else
-          debug("Rename #{relay_dir}")
-          cmd_result("mv #{relay_dir} #{relay_dir}_#{get_additional_property(ResetClusterCommand::ARCHIVE_LOGS_SUFFIX)}")
-          mkdir_if_absent(relay_dir)
+        if File.exist?(relay_dir)
+          if get_additional_property(ResetClusterCommand::DELETE_LOGS) == "true"
+            debug("Remove all files in #{relay_dir}")
+            cmd_result("rm -f #{relay_dir}/*")
+          else
+            debug("Rename #{relay_dir}")
+            cmd_result("mv #{relay_dir} #{relay_dir}_#{get_additional_property(ResetClusterCommand::ARCHIVE_LOGS_SUFFIX)}")
+            mkdir_if_absent(relay_dir)
+          end
         end
       }
     end
